migrate: replace 6 React contexts with route-scoped Jotai atoms
Replaces all six demo Contexts (Gaia, GravitationalLensing, SWEBench,
TrajectoryAtlas, Transformer, LatentInsights) with co-located atoms.ts
files using a route-scoped <Provider> from jotai. Net -2,286 LOC.

Each demo:
- Atoms split by re-render boundary (e.g. selectedMetric is its own atom,
  no longer flushes every consumer of the old monolithic context value).
- DuckDB init lives in a write-only initializeAtom triggered via
  atom.onMount (no separate Initializer component, no useRef guard —
  the if(get(coordinatorAtom)) return line handles StrictMode).
- vgplot Selection / Coordinator references held in atoms as opaque refs
  (set once at init; identity stability guarantees no cascading renders).
- Bridge components inside <Provider> port the old useEffects that watch
  derived inputs and dispatch async refreshes (Trajectory filter, Trajectory
  source switch, Transformer prompt/metric/brush refreshes, LatentInsights
  SSE cleanup).
- LatentInsights keeps its 11-event-type SSE reducer wholesale via
  atomWithReducer; only the SSE EventSource handle moved to its own atom
  for atomic teardown on Provider unmount.
- Removes the module-level `let coordinatorRef:any = null` anti-pattern
  in GravitationalLensing.

Performance wins: components now subscribe only to atoms they read, so
e.g. flipping selectedMetric in Transformer no longer re-renders TokenList
when only selectionStats changes. Same isolation across SWEBench's 7
consumers, TrajectoryAtlas's stats panel, etc.

Bumps jotai 2.12.4 → 2.20.0.
