+++ package.json
@@ -15,12 +15,12 @@
     "deploy": "gh-pages -d dist"
   },
   "dependencies": {
+    "@any_table/core": "^0.2.0",
+    "@any_table/react": "^0.2.0",
     "@chakra-ui/react": "^3.32.0",
     "@duckdb/duckdb-wasm": "^1.30.0",
     "@emotion/react": "^11.14.0",
     "@emotion/styled": "^11.14.0",
-    "@any_table/core": "^0.2.0",
-    "@any_table/react": "^0.2.0",
     "@kvis/packed-radial-tree": "file:/./src/lib/kvis-packed-radial-tree-1.0.0.tgz",
     "@mdx-js/react": "^3.1.0",
     "@mdx-js/rollup": "^3.1.0",
@@ -42,7 +42,7 @@
     "d3": "^7.9.0",
     "d3-scale-chromatic": "^3.1.0",
     "framer-motion": "^12.9.4",
-    "jotai": "^2.12.4",
+    "jotai": "^2.20.0",
     "next-themes": "^0.4.6",
     "prismjs": "^1.30.0",
     "react": "^18.3.1",
