+++ src/pages/Demos/LatentInsights/index.tsx
@@ -1,13 +1,25 @@
+import { useEffect } from "react";
+import { Provider, useSetAtom } from "jotai";
 import { Page } from "../../../components/Page";
-import { LatentInsightsProvider } from "../../../contexts/LatentInsightsContext";
 import { Dashboard } from "./components/Dashboard";
+import { cleanupSSEAtom } from "./atoms";
+
+/** Closes the SSE EventSource when the Provider unmounts (route change). */
+function SSECleanup() {
+  const cleanup = useSetAtom(cleanupSSEAtom);
+  useEffect(() => {
+    return () => cleanup();
+  }, [cleanup]);
+  return null;
+}
 
 export function LatentInsights() {
   return (
     <Page>
-      <LatentInsightsProvider>
+      <Provider>
+        <SSECleanup />
         <Dashboard />
-      </LatentInsightsProvider>
+      </Provider>
     </Page>
   );
 }
